Version 0.4.1
- Fix infinite loop to find window when easyhtml buffer is hidden. Thanks to Jonathon Merz who pointed out the bug and send me the patch.
'The drawback to this fix is that it will reinitialize all the buffer
mappings and such for the EasyHtml buffer every time the buffer is opened but
is not in a window already', he says. Hari Krishna Dara has made a great script named bufNwinUtils.vim which contains useful functions to handle buffers and windows. It will allow me to avoid the drawback. I'm interesting to use his script with mine, but it requires to load entire bufNwinUtils.vim for only  two or three functions used. What do you think about that ? give me feedback, thanks.